Round-Up Commit (See Description) - Get rid of this strange, non-standard package scheme.   - Refactored from RCMain, RCListeners, and RCCommands to me.viscar.townyrelationalcolors (.commands|.listeners)   - Respect package naming scheme of author's other projects - Expand maven to actually allow building   - Add properties section to store changing variables   - Define proper source and resource directories (src/main/java/ & src/main/resources/ respectively)   - Enable resource filtering - plugin.yml can now be updated by the pom without having to update variables in multiple locations. - Set source encoding to UTF-8 to prevent encoding issues across platforms - Reset api-version to 1.13 (update later if needed...) - Update Spigot API library to 1.16.4
