--- plugin.yml
@@ -1,12 +1,0 @@
-name: RelationColors
-main: RCMain.RelationColors
-Author: Viscar
-version: 1.0
-api-version: 1.15
-depend:
-  - Towny
-  - ProtocolLib
-softdepend:
-  - AnimatedNames
-commands:
-  rc:
