+++ pom.xml
@@ -4,18 +4,41 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
-    <artifactId>TownyRelationColors</artifactId>
+    <groupId>me.viscar.townyrelationalcolors</groupId>
+    <artifactId>TownyRelationalColors</artifactId>
     <version>1.0-SNAPSHOT</version>
+
+    <properties>
+        <java.version>1.8</java.version> <!--Valid: 1.8, 11, 14, 15-->
+        <project.bukkit.APIVersion>1.13</project.bukkit.APIVersion>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.maven.compiler.version>3.8.1</project.maven.compiler.version>
+        <project.maven.jdoc.version>3.2.0</project.maven.jdoc.version>
+    </properties>
+
     <build>
+        <defaultGoal>clean package</defaultGoal>
+        <sourceDirectory>src/main/java</sourceDirectory>
+        <resources>
+            <resource>
+                <directory>src/main/resources</directory>
+                <filtering>true</filtering>
+            </resource>
+        </resources>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>${project.maven.compiler.version}</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>${java.version}</source>
+                    <target>${java.version}</target>
                 </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-javadoc-plugin</artifactId>
+                <version>${project.maven.jdoc.version}</version>
             </plugin>
         </plugins>
     </build>
@@ -46,7 +69,7 @@
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
-            <version>1.15.2-R0.1-SNAPSHOT</version>
+            <version>1.16.4-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
