+++ src/main/java/me/viscar/townyrelationalcolors/Rank.java
@@ -1,4 +1,4 @@
-package RCMain;
+package me.viscar.townyrelationalcolors;
 
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
