+++ src/main/java/me/viscar/townyrelationalcolors/TownTeamsContainer.java
@@ -1,4 +1,4 @@
-package RCMain;
+package me.viscar.townyrelationalcolors;
 
 import com.comphenix.protocol.PacketType;
 import com.comphenix.protocol.ProtocolLibrary;
